Version1.9
This is a prototype script, deploy in a production environment at your own risk.

this is a simple script that will print TCP/IP IPv4 configurations and prompt the user if they would like to reset network configuration settings.

Future updates to include linking MAC addresses with IP addresses. 

Possible uses: Can be  deployed on a server in a local subnet and accessed via a management server via SHH to increase efficiency when updating the internal IP addressing table of a range of servers in a local subnet.